refactor: 패키지 이동, dto 및 메소드명 변경
