--- src/main/java/com/canvasdiary/canvasdiaryprototype/diary/DiaryCreateResponse.java
@@ -1,9 +1,0 @@
-package com.canvasdiary.canvasdiaryprototype.diary;
-
-public record DiaryCreateResponse(
-        String description,
-        String emotion,
-        String extractedEmotion,
-        String canvasImageUrl
-) {
-}
