+++ src/main/java/com/canvasdiary/canvasdiaryprototype/diary/dto/DiaryCreateRequest.java
@@ -1,4 +1,5 @@
-package com.canvasdiary.canvasdiaryprototype.diary;
+package com.canvasdiary.canvasdiaryprototype.diary.dto;
+
 import com.canvasdiary.canvasdiaryprototype.diary.canvas.Style;
 import com.canvasdiary.canvasdiaryprototype.diary.emotion.Emotion;
