+++ src/main/java/com/canvasdiary/canvasdiaryprototype/diary/dto/DiaryCreateViewResponse.java
@@ -1,0 +1,11 @@
+package com.canvasdiary.canvasdiaryprototype.diary.dto;
+
+import java.util.List;
+
+public record DiaryCreateViewResponse(
+        String description,
+        String emotion,
+        String extractedEmotion,
+        List<String> canvasImageUrl
+) {
+}
