+++ src/main/java/com/canvasdiary/canvasdiaryprototype/global/client/gemini/GeminiEmotionEmbedding.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum Emotion {
+public enum GeminiEmotionEmbedding {
 
     ANGER(new double[] {0.00036627866, -0.070826955, -0.015350174, -0.043797996, 0.023196205, 0.023519915, -0.024480592, 0.0064234766, 0.02441439, 0.049629334, 0.030453835, -0.0014627888, -0.021532241, -0.018007016, 0.0023313991, -0.050444845, 0.008838839, 0.0107716685, 0.0011984543, 0.003924937, 0.007839369, 0.021010676, -0.050415874, 0.026601896, -0.0074992613, 0.026475076, 0.07000654, -0.093186565, -0.0219551, 0.024688251, -0.03888582, 0.051990584, -0.055437233, -0.007198258, 0.00986914, -0.031851582, 0.008125888, 0.001369545, 0.042197168, 0.011453327, -0.008408682, -0.012614607, -0.05466123, -0.04115574, 0.03099183, -0.020245949, 0.011154014, -0.028916288, 0.013166551, -0.06778196, 0.049315345, -0.03311663, 0.045844574, -0.014235086, -0.042705644, -0.03780185, 0.07524046, 0.030692149, -0.035971645, -0.0071625034, 0.004489803, 0.04438943, -0.01942825, 0.061259586, 0.015423308, -0.027815174, 0.014523258, -0.012951834, 0.053262234, -0.018824384, 0.015905444, 0.0053603505, 0.06502664, -0.03663673, -0.000536316, -0.07512496, -0.008461147, 0.046909735, 0.0023172316, 0.07527665, 0.013250037, -0.10255497, -0.012079425, -0.015456549, -0.027341904, -0.015065761, -0.0029533608, 0.040104993, 0.045160845, 0.036126453, -0.04800378, -0.031076973, -0.006053585, -0.08316381, -0.0013084664, 0.034334686, -0.020627504, 0.036824778, -0.021632465, -0.037743386, 0.003127868, -0.038079314, -0.006751566, -0.013256466, 0.01715441, 0.020737648, 0.03711938, 0.06570432, 0.035107743, -0.00028474248, -0.014024044, 0.010321317, -0.0061357743, -0.03289243, 0.016288718, 0.056814473, 0.0014724412, 0.08316739, 0.024183493, 0.058991153, -0.0020632243, -0.0519902, 0.049253143, -0.045170475, -0.017677769, 0.015676362, -0.001243756, 0.025943095, 0.04540258, 0.047376335, -0.0370189, -0.0010816281, -0.0006255375, 0.05843728, 0.08139005, 0.051482376, 0.017953806, -0.026730947, 0.056118473, 0.011492796, 0.0026686264, 0.012195242, 0.021669408, 0.012182083, -0.009110708, 0.04755686, -0.064097896, -0.017561123, 0.017212568, -0.00448884, 0.005710719, 0.042429462, -0.0536028, -0.007989118, 0.05626224, -0.0060325386, -0.01716225, -0.01309776, -0.013646949, -0.02908075, -0.009856284, 0.033497177, -0.0029582237, 0.036787655, -0.041603442, -0.0367336, 0.027779534, 0.0028387064, 0.026822662, 0.016367579, -0.02459561, 0.0044496455, -0.058905717, -0.060202748, -0.004818778, -0.02960869, -0.058847178, 0.007159367, -0.058182262, 0.0011632034, -0.0081250565, -0.0346072, 0.008561833, 0.025371432, 0.041548587, -0.0037040454, 0.01815437, -0.03971953, -0.01891851, 0.011930184, -0.011968631, -0.028499562, -0.010823559, -0.07422832, -0.030227272, -0.044783514, -0.0005798821, 0.009777118, 0.022088513, -0.047930967, 0.012615071, 0.074946284, -0.014836287, 0.0008344985, 0.036872156, -0.028395869, 0.03415926, -0.033784796, 0.012334977, 0.042412598, -0.041421253, -0.0017775539, -0.038343232, 0.032147482, 0.030476496, -0.0063250195, 0.00074940355, 0.031892806, 0.010571302, -0.025104186, -0.030788343, -0.0020762696, -0.038248807, 0.022981577, 0.011764523, 0.06592368, 0.042734496, -0.02066992, 0.028799443, -0.08539295, -0.0071601695, 0.052230325, 0.01265585, 0.0018700759, 0.08613652, -0.019855779, -0.010222885, -0.010286552, 0.04373864, 0.030861871, -0.04619367, 0.055889502, 0.04340073, 0.044392146, -0.037495647, 0.005090721, 0.01506533, 0.041322954, 0.007486256, 0.04296941, -0.018776033, -0.018696912, -0.0024843907, 0.013386704, -0.11077184, 0.028462365, -0.06661907, -0.028423872, -0.004566844, -0.0028030432, 0.04717277, -0.007898089, 0.0071629826, 0.009037111, -0.014201927, -0.009372889, -0.0039415476, -0.018675627, -0.017289538, 0.045222137, -0.027180947, -0.030669333, 0.04094874, -0.000707994, 0.01828456, 0.0023999505, 0.0064469352, 0.04536496, 0.024582397, 0.0053655454, -0.00069232215, -0.03851905, 0.04570247, -0.06110789, -0.025725618, -0.03228014, -0.057874475, -0.040220786, 0.00048566097, -0.028252043, -0.03139933, 0.012081332, -0.041362155, -0.038335554, -0.03016004, -0.044345014, -0.02660439, 0.033526555, 0.008514456, -2.8369603e-05, -0.003399602, -0.06523597, 0.01015318, -0.08272147, 0.007223889, 0.012541745, -0.008114546, 0.023005718, 0.02921668, -0.005422916, 0.014326773, -0.0041157925, -0.052559778, 0.020153768, 0.024670392, 0.03596248, 2.4706673e-05, 0.0145351775, -0.070160225, 0.019806698, 0.057887156, 0.053325098, 0.021829301, -0.015207407, 0.0047318395, 0.041621797, -0.055997744, 0.053662945, 0.00622129, -0.021613648, 0.0029282833, -0.021742161, -0.038316276, 0.057124015, -0.020590821, 0.011860318, -0.043448575, 0.005260626, -0.0072159213, 0.0039443313, 0.016226681, 0.0021318241, -0.023943102, -0.08604558, 0.0008120858, -0.039049022, -0.008363761, 0.02628159, 0.055752564, 0.07410372, -0.0021166604, 0.03214227, -0.020387668, -0.020220175, 0.025433151, 0.0044751014, 0.032395463, 0.016247118, 0.06127831, -0.035189264, -0.052063555, 0.07022598, -0.0063727675, 0.033927463, 0.032513026, 0.01648868, 0.034000844, 0.022902275, -0.024939489, -0.00051005173, -0.028008835, -0.028924726, -0.008930342, -0.025771832, -0.01370661, -0.05077857, -0.016223047, -0.016956482, 0.019485679, -0.015628234, -0.01744789, -0.020865042, 0.035279028, 0.06776303, 0.021225186, 0.006110669, 0.019125368, 0.04663303, 0.009203263, 0.02888032, 0.04250187, 0.055643916, 0.055443816, 0.030713355, -0.0051422087, 0.04059326, 0.024263693, -0.07923062, 0.07068526, -0.04018101, -0.013312973, -0.01160726, -0.001266306, -0.021042293, -0.034808684, 0.0045950096, 0.0165496, -0.052391738, 0.008973441, -0.0063516097, 0.023005923, 0.010372779, 0.041151483, -0.016569387, -0.03614827, -0.02507301, 0.06058971, -0.005252807, -0.04372291, 0.031292547, -0.03885335, -0.02206465, 0.02865736, -0.02151582, -0.028317668, -0.010497701, 0.008447014, -0.037150484, -0.06835301, -0.005727414, 0.020089554, 0.04136245, -0.03819462, 0.02998397, -0.0015510026, -0.09094654, -0.015775256, 0.013768867, -0.013604095, -0.031829588, 0.0012299559, -0.045242418, 0.0004171359, 0.013450703, -0.09337715, -0.00647558, -0.012621617, -0.03415363, 0.0022020938, -0.05672877, 0.03737742, -0.08756255, 0.005848537, -0.07041305, 0.017238822, -0.02495426, -0.042252343, 0.046204433, -0.0334503, 0.006890448, -0.06132209, -0.037111502, -0.011496106, -0.02755034, 0.013716472, -0.022373602, 0.016105276, 0.0051255375, 0.008145512, 0.0062138736, 0.00090743747, -0.034672257, -0.03743107, 0.008553682, 0.0062694736, -0.02701366, -0.06051431, -0.03570312, 0.016075578, -0.02179524, 0.05199612, -0.010619983, 0.06354413, 0.04190532, 0.010522639, 0.0046481914, 0.034938496, -0.044225357, -0.028490853, 0.045840282, -0.023478908, -0.055454668, -0.050402474, -0.057030637, -0.015660958, 0.032021824, -0.038512524, 0.027550492, 0.036034074, 0.002936173, 0.0039389506, -0.034227245, -0.016679762, -0.022597551, 0.047461413, -0.07864558, -0.009593177, -0.029661845, 0.00028480554, -0.05696323, -0.001134428, 0.040024616, -0.028317211, 0.0076065008, 0.04985905, -0.041495275, -0.0020048448, 0.024841525, 0.020669678, 0.03709539, 0.06731936, -0.0038912944, -0.09802012, -0.013767556, -0.02293593, -0.05266017, -0.011090115, 0.05173256, -0.054415137, -0.008743669, -0.058007922, 0.053872243, -0.08034007, -0.025928188, -0.052180354, -0.02183362, 0.04006225, -0.03610565, 0.026061777, -0.026127858, 0.03432089, -0.026932344, 0.03308791, 0.032242972, 0.002574761, 0.005941122, 0.055098526, -0.08268066, -0.0038781643, -0.008928034, -0.003415907, -0.020166619, 0.022907466, -0.045946684, 0.01414059, 0.05561238, 0.009385163, -0.024642257, 0.02190973, 0.031282816, -0.024743453, -0.036122654, -0.00691566, -0.10082132, 0.0882347, -0.0010043085, -0.012902406, -0.01087235, 0.055642266, -0.00014122132, 0.009511363, 0.03792847, 0.046441637, -0.0052278126, 0.011284871, -0.014038459, -0.04495369, -0.013708433, -0.032127574, -0.02095313, 0.015376052, -0.06557728, 0.011147531, 0.031508915, -0.052457403, 0.054762162, -0.030640509, -0.009815483, -0.0040262416, 0.014379875, -0.061504368, 0.0010636183, -0.054109823, 0.07647969, -0.050639573, 0.0028312067, 0.047082175, 0.0064740507, -0.013575361, -0.03616612, -0.011174286, -0.053276613, 0.06779609, -0.07762937, 0.043703694, -0.013457372, 0.006191658, 0.00095835654, -0.017992109, 0.032170825, -0.020408852, -0.052687492, -0.012091451, 0.013461492, -0.04431172, 0.008842752, 0.051579084, 0.011402583, -0.062043067, -0.02904613, 0.02826587, -0.07519633, 0.029902954, 0.0366335, 0.07322498, -0.04899892, 0.034200367, -0.032260638, 0.028673714, 0.03861677, 0.025365517, 0.037335407, 0.02672696, 0.04168721, -0.04792213, 0.03214155, 0.03798889, 0.01134968, 0.00085469615, 0.03534652, -0.08665209, -0.0071371915, 0.044856407, 0.0002035114, -0.031494364, 0.03719543, 0.028783275, -0.10800851, -0.0335873, 0.02556313, -0.04083057, 0.02245558, -0.021922207, 0.007271506, 0.025534576, 0.010040937, -0.07080663, -0.020304326, 0.016611898, -0.017389987, -0.030739406, -0.035696004, -0.014386981, -0.010750457, -0.029067468, -0.0036931883, -0.062219743, -0.052396797, 0.0050548147, -0.00015333918, -0.053244356, 0.058605522, 0.0493728, 0.012186038, 0.05279283, 0.044327028, 0.008919091, -0.024836095, -0.060059693, 0.04022796, -0.020947795, -0.018613854, -0.014054277, -0.009829306, -0.019480916, 0.04844231, 0.010936403, -0.032762505, -0.014891943, -0.029817311, 0.05121774, -0.014853717, -0.057674795, -0.03238302, 0.047714986, -0.009271314, -0.023387924, 0.020175342, -0.0052558216, 0.039583907, -0.018010681, -0.009561307, -0.0074558672, -0.0055476357, 0.015406492, 0.03170371, -0.030912954, -0.04627847, 0.024171658, 0.017257916, 0.016329389, -0.0022346817, 0.0035314264, 0.04551409, 0.02333835, -0.04517144, -0.009169872, -0.021008266, -0.066322364, 0.029807383, -0.02311807, 0.024333391, 0.050832197, -0.021675788, -0.004018456, 0.030785225, -0.07268611, 0.019887635, -0.04567573, -0.07464042, 0.015230038, -0.050599165, -0.014454001, -0.0070348685, -0.018249985, 0.062161844, -0.03466089, 0.016534988, -0.03927901, -0.03624997, -0.048068862, 0.007649897, 0.028124463, 0.047968537, 0.027744222, -0.037469327, 0.028213667, -0.028962187, -0.025236731, -0.0014608477, -0.02435135, -0.026030844, 0.023806376, 0.08446295, 0.03518877, -0.010194815, 0.039760996, -0.009162086, 0.0139915, -0.009100025, -0.071320966, 0.04070641, -0.02948992, 0.009223466, -0.025310347, -0.010154387, -0.033918295, 0.024861684}),
     FEAR(new double[] {0.0073998775, -0.060280066, -0.0027088234, -0.055427406, 0.028010594, 0.048781257, 0.020949882, 0.018535974, 0.04399848, 0.046217535, 0.021266194, -0.015999172, -0.009092171, -0.021703862, 0.0071101165, -0.03344851, 0.031223916, 0.022053942, 0.016775895, -0.012150517, 0.014075205, 0.022569496, -0.00043881286, 0.01614666, -0.0070683584, 0.031731267, 0.05857133, -0.0726531, -0.010245238, 0.064418584, -0.060968764, 0.013065976, -0.04547665, -0.001094074, 0.012260322, -0.053829532, -0.014748812, 0.06278493, 0.024380261, 0.003898687, -0.0072190915, -0.0054331305, -0.045746487, -0.01887811, 0.011831253, -0.042903375, 0.029261211, -0.038913637, -0.051896814, -0.0746009, 0.05595062, -0.0269856, 0.06144897, 0.020143993, -0.024395006, -0.045066066, 0.051990658, 0.0020049661, -0.050812323, -0.03050071, 0.042788256, 0.049445428, -0.040637355, 0.03324475, -0.023249544, 0.021354828, 0.003291545, 0.060084965, 0.060994662, -0.051608954, 0.018937387, -0.004491766, 0.013319996, -0.023824228, -0.019857975, -0.038680334, -0.032579765, 0.054918133, -0.0075018373, 0.041581165, 0.021549447, -0.05853418, -0.014844343, -0.057044283, -0.04311988, -0.035530984, -0.041088782, 0.030990863, -0.0044260966, 0.04191784, -0.02973165, -0.037268393, -0.00042382826, -0.05385226, -0.025052657, 0.0360679, -0.019205347, 0.012428302, 0.011771351, -0.03285932, -0.013387608, -0.053112548, -0.028844591, -0.02588108, 0.00018975344, 0.019057687, 0.021018244, 0.010731631, 0.0018349302, 0.009850976, -0.057771508, 0.026031796, -0.027797222, -0.08167643, 0.048925415, 0.008703614, 0.034083758, 0.07847017, 0.04392419, 0.084317505, 0.020305311, -0.022454634, 0.05225386, -0.0110187875, -0.036542285, -0.007970311, 0.0077922908, 0.0069910698, 0.040417142, 0.019265374, -0.025584985, -0.013563588, 0.008094332, -0.026551759, 0.05639308, 0.087789975, 0.037941, -0.02693312, 0.049147133, 0.00827871, 0.023309277, 0.011274275, 0.010498787, -0.0010456896, -0.01864513, 0.03143673, -0.06248022, -0.017765595, 0.0263784, 0.0005090219, 0.02059245, 0.0097078765, -0.056048922, -0.0016526232, 0.08313541, -0.013785181, -0.030735692, 0.034647435, 0.009774065, -0.031457834, 0.0025449367, 0.004849039, 0.01694812, 0.023892744, -0.0066086417, -0.027076421, -0.022723535, 0.008505719, 0.024388846, 0.03947794, 0.010743861, 0.055519644, -0.048506945, -0.08218866, -0.026648935, -0.07411652, -0.03554325, -0.014516554, -0.039506786, -0.03670101, -0.015201126, -0.053242464, 0.004457664, 0.014337788, 0.049906667, 0.011656081, 0.017904093, -0.063751325, -0.014529079, -0.011000245, 0.011627452, 0.00019760632, -0.022705214, -0.026041528, 0.01821799, -0.0147583885, -0.016352154, 0.0011641338, 0.020730218, -0.019831404, -0.0038091156, 0.10329318, -0.0075338823, 0.00091830734, 0.061434187, -0.014442858, 0.037389796, -0.036684662, 0.01953009, 0.0191362, -0.067809395, -0.008431193, -0.048888806, 0.03259712, 0.012771837, -0.0070370333, 0.0006991707, 0.05517485, -0.00019261423, -0.05139131, -0.020342993, -0.015378644, -0.036955174, 0.012377845, 0.0039665955, 0.061651126, 0.019917104, 0.009391361, 0.003123251, -0.039821133, -0.012461606, 0.032082725, 0.008460567, -0.036762808, 0.0776158, -0.030890778, 0.0038773827, -0.02601717, 0.016525216, 0.027728213, -0.041036464, 0.013791393, 0.04437937, 0.01129718, -0.06371095, 0.046655998, -0.037537817, 0.011684741, -0.040247295, 0.030473985, -0.0009252435, 0.009581151, 0.024967302, 0.018133657, -0.07794783, 0.039700672, -0.055657767, -0.021320285, 0.03287362, 0.0054628206, 0.05317667, -0.010029198, -0.001989656, -0.001909868, -0.021067774, 0.020331822, -0.011034423, -0.0013427791, 0.0039869216, 0.036317643, 0.027415855, -0.047755573, 0.03403023, -0.022947337, -0.0240577, -0.0009994749, 0.0051731155, 0.026500953, 0.038943615, 0.018366495, -0.02037214, -0.021476282, 0.027567895, -0.0451173, 0.006466073, -0.057595994, -0.049876995, -0.0035964234, 0.009858282, -0.040300235, -0.047187842, 0.0071316953, -0.0320864, -0.03367463, -0.027637785, 0.0036579012, 0.019815575, 0.018715907, 0.03065154, -0.0026208677, 0.02414028, -0.067984626, -0.009211629, -0.049135447, 0.021342253, 0.053930104, 0.01408795, 0.002868026, 0.022593413, -0.0038512747, -0.009933177, -0.00048457657, -0.0587623, 0.02177184, 0.049488142, 0.01677354, 0.011868904, 0.033758197, -0.052291624, 0.024915334, 0.045634627, 0.068806335, 0.051111355, 0.034092862, 0.033331245, 0.01523768, -0.05976093, 0.06088426, -0.034642085, -0.009366958, 0.002757303, 0.008448278, -0.02498332, 0.053524304, -0.0068641966, 0.011914409, -0.04628068, 0.003217543, -0.037589107, 0.007518356, 0.03673847, 5.6369605e-05, -0.0127552515, -0.10242028, 0.008515535, 0.0015350233, 0.009099362, 0.03094879, 0.050140098, 0.043894045, -0.01284332, 0.04886674, -0.027570983, -0.013052005, 0.017825712, -0.046290983, 0.023986908, 0.038395196, 0.0005694912, -0.04194408, 0.010627039, 0.08117465, 0.013018764, 0.034471843, 0.008035837, 0.025837146, 0.020325758, 0.01948273, -0.01984616, 0.031828336, -0.028986337, -0.033734348, -0.004036412, -0.021716425, 0.016130848, -0.053735558, -0.013909548, -0.0036208197, 0.013244635, -0.0004963354, 0.0005244845, -0.03910875, 0.056557447, 0.040149525, 0.04040249, 0.021376848, 0.014858773, 0.048406452, 0.005394406, 0.012515416, 0.052707337, 0.018628677, 0.06550349, 0.044419922, -0.039291598, 0.02347703, 0.016508155, -0.052690007, 0.046097193, -0.043649316, -0.009736973, -0.011916252, -0.031520393, -0.02054201, -0.035472877, -0.027029453, 0.017329458, -0.06072937, -0.0068317917, -0.0023477073, 3.4724926e-05, 0.041242186, 0.0006321989, -0.038520187, -0.028744973, 0.0024305084, 0.036346022, -0.000209839, -0.013412066, 0.029695729, -0.029531097, 0.01072106, -0.025475908, 0.019145736, -0.015008592, -0.037260756, -0.004395826, 0.025991559, -0.032470755, -0.0038910478, 0.051344153, 0.05418623, -0.038908843, 0.02749107, 0.011047113, -0.043650486, -0.01322488, -0.0046757176, -0.031739287, -0.020902844, 0.0049646725, -0.027292851, 0.035776734, 0.027662883, -0.09457237, -0.0032800839, -0.03572191, -0.04249659, 0.016879665, -0.07373088, 0.024280546, -0.111305684, 0.029446384, -0.078111246, 0.029101314, 0.0032720445, -0.033907156, 0.04975285, 0.010853832, -0.0016918149, -0.05378345, 0.0074897353, 0.0016483939, -0.03242231, 0.030256592, -0.027604528, 0.02781093, -0.010420249, 0.009240871, 0.025085071, 0.0036384223, -0.051492274, 0.00050659716, 0.00599467, -0.010192075, 0.009227056, -0.054423254, -0.01915563, 0.0077426117, -0.026643686, -0.013347921, -0.0017228493, 0.0113774575, 0.02082715, 0.040525883, -0.029139394, 0.002823454, -0.036843136, 0.014370143, 0.089597814, -0.017457211, -0.015926573, -0.02700918, -0.05823912, -0.005950321, 0.04329004, -0.042775627, 0.03059151, 0.029767184, 0.015286593, -0.002950189, -0.043853134, -0.022371585, -0.01994723, 0.06612876, -0.089240395, -0.006063089, -0.0034885406, 0.009368573, 0.0025102585, 0.0095092505, 0.043055754, -0.025129355, -0.008182675, 0.067736484, -0.047166873, 0.008421238, -0.0013772849, 0.041336585, 0.008704954, 0.06940473, -0.0074202465, -0.0991402, -0.03895212, -0.007223581, -0.022333497, -0.015374198, 0.04835606, -0.0773339, -0.009055652, -0.0471559, 0.07645649, -0.06957246, 0.020960703, -0.018198475, 0.025479246, 0.04451824, 0.0041134716, 0.005512022, -0.007822187, 0.039617307, -0.041107398, 0.012981368, 0.05485634, -0.03239026, -0.00490134, 0.01976878, -0.09184662, 0.00025589784, -0.02393914, -0.020177063, -0.024437517, 0.021864016, -0.024556642, -0.0048355046, 0.034627397, 0.020445492, -0.029875573, 0.027674187, -0.0126602575, -0.023823177, -0.025792196, -0.015611297, -0.07604322, 0.082649924, 0.009183887, -0.04000815, -0.027357062, 0.08571113, 0.015393489, -0.019154865, 0.06291511, 0.036623783, -0.013012574, 0.030888, -0.0024489479, -0.044650897, -0.0006127221, -0.027496586, 0.0028529183, 0.0133861, -0.08989483, 0.0070748045, 0.07720992, -0.08198565, 0.07767964, 0.0021112596, -0.008110034, 0.0043245917, -0.00252213, -0.062891014, -0.005499769, -0.013305509, 0.05375116, -0.042025447, 0.0028535414, 0.038735032, -0.004829643, -0.007359332, -0.018448427, 0.00067585864, -0.061171323, 0.08167967, -0.06375546, 0.06191301, -0.04518729, 0.0004523999, 0.03551318, -0.017942116, 0.051569328, -0.037960984, -0.07568002, 0.032772254, 0.009651151, -0.067787595, 0.018289108, 0.04462833, 0.02498332, -0.047386117, -0.016769642, 0.008314387, -0.057089638, 0.0497095, 0.024537908, 0.06972814, -0.0196796, 0.01701028, -0.019481292, 0.050915774, 0.06635185, 0.050687924, 0.04395239, -0.009056635, -0.004381125, -0.029105432, 0.04078117, 0.05472243, 0.03823235, 0.026467057, -0.031926557, -0.06431147, 0.0011236144, 0.053542763, -0.010638033, -0.008985005, 0.04414878, 0.0024654854, -0.0615717, -0.013681704, 0.009312316, -0.0060716295, 0.0032774075, 0.023891859, 0.005124223, -7.5020245e-05, 0.042276923, -0.051514205, -0.043647893, 0.0056514293, 0.011922157, -0.0016548226, -0.013031414, 0.016975533, -0.020812852, -0.021237891, -0.033294395, -0.057690915, -0.0703719, -0.0046332367, 0.041385658, -0.07938435, 0.057268344, 0.027036391, 0.043093555, 0.03182319, 0.049825262, -0.0122067155, 0.041239783, -0.016381817, 0.028723184, -0.0007570719, -0.027815657, -0.03240804, 0.028128194, -0.04264292, 0.024110196, 0.042190462, -0.0050559114, 0.010878467, -0.005897686, 0.0341669, -0.013028465, -0.011780748, 0.0122978315, 0.04387685, 0.0037825974, 0.018710501, 0.028650915, 0.015274165, 0.037146866, -0.0005589193, -0.032803614, -0.043696124, 0.0058261217, -0.0018822306, 0.018156653, -0.020489108, -0.014593249, 0.06372961, 0.008079241, 0.007394446, -0.035641506, 0.012092773, 0.0241119, 0.03318791, -0.051731553, 0.074888445, -0.015168175, -0.037737805, 0.027117172, 0.02116624, 0.013795254, 0.056236763, -0.01134726, -0.003896061, 0.032792058, -0.069188766, 0.03102121, -0.03701853, -0.075194985, -0.0048495326, -0.059412792, -0.03251438, 0.0029890006, -0.030963155, 0.12363498, -0.010264563, 0.020822542, -0.034567162, -0.03172084, -0.058890957, 0.0064984635, 0.015973786, 0.06667226, 0.008059857, 0.001868991, 0.0320169, -0.044734783, -0.02926044, -0.0022881317, -0.04382645, 0.013303137, 0.03369169, 0.06891315, 0.058421463, 0.012598782, 0.029525423, 0.014838216, -0.017384896, -0.008099139, -0.06098322, 0.018434511, -0.0008080798, 0.027162956, 0.009267496, 0.009067387, -0.039800752, -0.003985103}),
